+++ test/bootstrap_form_test.rb
@@ -7,7 +7,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "default-style forms" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
       </form>
     HTML
@@ -16,7 +16,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "default-style form fields layout horizontal" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group row">
           <label class="col-form-label col-sm-2 required" for="user_email">Email</label>
@@ -68,7 +68,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "default-style form fields layout inline" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group form-inline">
           <label class="mr-sm-2 required" for="user_email">Email</label>
@@ -120,7 +120,7 @@ class BootstrapFormTest < ActionView::TestCase
     # generate a default ID for the form element.
     test "default-style forms bootstrap_form_with Rails 5.1+" do
       expected = <<-HTML.strip_heredoc
-        <form accept-charset="UTF-8" action="/users" data-remote="true" method="post" role="form">
+        <form accept-charset="UTF-8" action="/users" data-remote="true" method="post">
           #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         </form>
       HTML
@@ -130,7 +130,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "inline-style forms" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user form-inline" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user form-inline" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group">
           <label class="mr-sm-2 required" for="user_email">Email</label>
@@ -175,7 +175,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "horizontal-style forms" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group row">
           <label class="col-form-label col-sm-2 required" for="user_email">Email</label>
@@ -226,7 +226,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "horizontal-style form fields layout default" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group">
           <label class="required" for="user_email">Email</label>
@@ -272,7 +272,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "horizontal-style form fields layout inline" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group form-inline">
           <label class="mr-sm-2 required" for="user_email">Email</label>
@@ -318,7 +318,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "existing styles aren't clobbered when specifying a form style" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="my-style" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="my-style" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group row">
           <label class="col-form-label col-sm-2 required" for="user_email">Email</label>
@@ -332,7 +332,7 @@ class BootstrapFormTest < ActionView::TestCase
                           bootstrap_form_for(@user, layout: :horizontal, html: { class: "my-style" }) { |f| f.email_field :email }
   end
 
-  test "given role attribute should not be covered by default role attribute" do
+  test "given role attribute be applied" do
     expected = <<-HTML.strip_heredoc
       <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="not-a-form">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
@@ -343,7 +343,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "bootstrap_form_tag acts like a form tag" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group">
           <label for="email">Your Email</label>
@@ -357,7 +357,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "bootstrap_form_for does not clobber custom options" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group">
           <label class="required" for="ID">Email</label>
@@ -370,7 +370,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "bootstrap_form_tag does not clobber custom options" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group">
           <label for="ID">Email</label>
@@ -390,7 +390,7 @@ class BootstrapFormTest < ActionView::TestCase
       name = "[misc]"
     end
     expected = <<-HTML.strip_heredoc
-    <form accept-charset="UTF-8" action="/users" method="post" role="form">
+    <form accept-charset="UTF-8" action="/users" method="post">
       #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
       <div class="form-check">
         <input class="form-check-input" id="#{id}" name="#{name}" type="checkbox" value="1" />
@@ -407,7 +407,7 @@ class BootstrapFormTest < ActionView::TestCase
     assert @user.invalid?
 
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group">
           <label class="required text-danger" for="user_email">Email can't be blank, is too short (minimum is 5 characters)</label>
@@ -423,7 +423,7 @@ class BootstrapFormTest < ActionView::TestCase
     assert @user.invalid?
 
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group">
           <label class="required text-danger" for="user_email">Email can&#39;t be blank, is too short (minimum is 5 characters)</label>
@@ -443,7 +443,7 @@ class BootstrapFormTest < ActionView::TestCase
       assert @user.invalid?
 
       expected = <<-HTML.strip_heredoc
-        <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+        <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
           #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
           <div class="form-group">
             <label class="required text-danger" for="user_email">Your e-mail address can&#39;t be blank, is too short (minimum is 5 characters)</label>
@@ -501,7 +501,7 @@ class BootstrapFormTest < ActionView::TestCase
     end
 
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="alert alert-danger">
           <p>Please fix the following errors:</p>
@@ -525,7 +525,7 @@ class BootstrapFormTest < ActionView::TestCase
     end
 
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="alert alert-danger">
           <p>Please fix the following errors:</p>
@@ -544,7 +544,7 @@ class BootstrapFormTest < ActionView::TestCase
     end
 
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="alert alert-danger">
           <p>Please fix the following errors:</p>
@@ -592,7 +592,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "custom label width for horizontal forms" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group row">
           <label class="col-form-label col-sm-1 required" for="user_email">Email</label>
@@ -608,7 +608,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "offset for form group without label respects label width for horizontal forms" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group row">
           <div class="col-md-10 offset-md-2">
@@ -624,7 +624,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "offset for form group without label respects multiple label widths for horizontal forms" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group row">
           <div class="col-sm-8 col-md-10 offset-sm-4 offset-md-2">
@@ -640,7 +640,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "custom input width for horizontal forms" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group row">
           <label class="col-form-label col-sm-2 required" for="user_email">Email</label>
@@ -656,7 +656,7 @@ class BootstrapFormTest < ActionView::TestCase
 
   test "additional input col class" do
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group row">
           <label class="col-form-label col-sm-2 required" for="user_email">Email</label>
@@ -680,7 +680,7 @@ class BootstrapFormTest < ActionView::TestCase
     end
 
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group">
           <label class="required" for="user_email">Email</label>
@@ -701,7 +701,7 @@ class BootstrapFormTest < ActionView::TestCase
     end
 
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group">
           <div class="field_with_errors">
@@ -726,7 +726,7 @@ class BootstrapFormTest < ActionView::TestCase
     end
 
     expected = <<-HTML.strip_heredoc
-      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+      <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
         #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
         <div class="form-group">
           <label class="required" for="user_email">Email</label>
@@ -747,7 +747,7 @@ class BootstrapFormTest < ActionView::TestCase
       end
 
       expected = <<-HTML.strip_heredoc
-        <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post" role="form">
+        <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post">
           #{'<input name="utf8" type="hidden" value="&#x2713;"/>' unless ::Rails::VERSION::STRING >= '6'}
           <div class="form-group">
             <label class="required" for="user_email">Email</label>